Move calculator to flow directory - calculate-your-child-maintenance
This calculator is specific to the calculate-your-child-maintenance flow and so
should be co-located.

* SmartAnswer::Calculators:ChildMaintenanceCalculator
  -> SmartAnswer::CalculateYourChildMaintenanceFlow::Calculator

* SmartAnswer::Calculators::ChildMaintenanceCalculatorTest
  -> SmartAnswer::CalculateYourChildMaintenanceFlow::CalculatorTest

Since these changes do not affect the behaviour and all the regression tests
pass, I've updated the checksums with the new location of the calculator.
